Adding template directory
