+++ Package.swift
@@ -13,10 +13,12 @@ let package = Package(
     ],
     targets: [
         .target(
-            name: "SwaggerSwift", dependencies: [
+            name: "SwaggerSwift",
+            dependencies: [
                 .product(name: "ArgumentParser", package: "swift-argument-parser"),
                 .product(name: "SwaggerSwiftML", package: "SwaggerSwiftML"),
-            ]),
+            ],
+            resources: [.copy("Templates")]),
         .testTarget(
             name: "SwaggerSwiftTests",
             dependencies: ["SwaggerSwift"]),
