+++ Sources/SwaggerSwift/Keywords.swift
@@ -1,10 +1,3 @@
-//
-//  File.swift
-//  
-//
-//  Created by Maximilian Scheid on 30/08/2021.
-//
-
 import Foundation
 
 let swiftKeywords: [String] = ["defer", "default"]
