+++ Sources/SwaggerSwift/Templates/DummyTest.swift
@@ -1,0 +1,7 @@
+import XCTest
+
+public struct DummyTest {
+    func testNetwork() {
+        XCTAssertTrue(true)
+    }
+}
