+++ Sources/SwaggerSwift/Templates/NetworkInterceptor.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+public protocol NetworkInterceptor {
+    func networkWillPerformRequest(_ request: URLRequest) -> URLRequest
+    func networkDidPerformRequest(urlRequest: URLRequest, urlResponse: URLResponse?, data: Data?, error: Error?) -> Bool
+}
