+++ Sources/SwaggerSwift/Templates/ParsingError.swift
@@ -1,0 +1,5 @@
+import Foundation
+
+public enum JSONParsingError: Error {
+    case invalidDate(String)
+}
