+++ Sources/SwaggerSwift/Templates/ServiceError.swift
@@ -1,0 +1,8 @@
+public enum ServiceError<ErrorType>: Error {
+    // An error occured that is caused by the client app, and not the request
+    case clientError(reason: String)
+    // The request failed, e.g. timeout
+    case requestFailed(error: Error)
+    // The backend returned an error, e.g. a 500 Internal Server Error, 403 Unauthorized
+    case backendError(error: ErrorType)
+}
