+++ Sources/SwaggerSwift/Templates/URLQueryExtension.swift
@@ -1,0 +1,15 @@
+import Foundation
+
+extension URLQueryItem {
+    init(name: String, value: Bool) {
+        self.init(name: name, value: value ? "true" : "false")
+    }
+
+    init(name: String, value: Int) {
+        self.init(name: name, value: String(value))
+    }
+
+    init(name: String, value: Int64) {
+        self.init(name: name, value: String(value))
+    }
+}
